Part 2

+++ Test/Models/CalculateInfo.cs
@@ -7,8 +7,8 @@ namespace Test.Models
 {
     public class CalculateInfo
     {
-        public double Field1 { get; set; }
-        public double Field2 { get; set; }
+        public decimal Field1 { get; set; }
+        public decimal Field2 { get; set; }
         public int Operation { get; set; }
 
     }
